requirements.txt
discord.py ~= 2.1
python-dotenv ~= 0.21
loguru ~= 0.6
psycopg ~= 3.1
jurigged ~= 0.5